+++ cvat-core/src/server-proxy.js
@@ -48,15 +48,17 @@
     async function chunkUpload(file, uploadConfig) {
         const params = enableOrganization();
         const {
-            endpoint, chunkSize, totalSize, onUpdate,
+            endpoint, chunkSize, totalSize, onUpdate, metadata,
         } = uploadConfig;
-        let { totalSentSize } = uploadConfig;
+        const { totalSentSize } = uploadConfig;
+        const uploadResult = { totalSentSize };
         return new Promise((resolve, reject) => {
             const upload = new tus.Upload(file, {
                 endpoint,
                 metadata: {
                     filename: file.name,
                     filetype: file.type,
+                    ...metadata,
                 },
                 headers: {
                     Authorization: Axios.defaults.headers.common.Authorization,
@@ -79,9 +81,13 @@
                         onUpdate(percentage);
                     }
                 },
+                onAfterResponse(request, response) {
+                    const uploadFilename = response.getHeader('Upload-Filename');
+                    if (uploadFilename) uploadResult.filename = uploadFilename;
+                },
                 onSuccess() {
-                    if (totalSentSize) totalSentSize += file.size;
-                    resolve(totalSentSize);
+                    if (totalSentSize) uploadResult.totalSentSize += file.size;
+                    resolve(uploadResult);
                 },
             });
             upload.start();
@@ -705,20 +711,39 @@
                 // keep current default params to 'freeze" them during this request
                 const params = enableOrganization();
 
-                let taskData = new FormData();
-                taskData.append('task_file', file);
+                const taskData = new FormData();
+                const uploadConfig = {
+                    chunkSize: config.uploadChunkSize * 1024 * 1024,
+                    endpoint: `${origin}${backendAPI}/tasks/backup/`,
+                    totalSentSize: 0,
+                    totalSize: file.size,
+                };
+
+                const url = `${backendAPI}/tasks/backup`;
+                await Axios.post(url,
+                    new FormData(), {
+                        params,
+                        proxy: config.proxy,
+                        headers: { 'Upload-Start': true },
+                    });
+                const { filename } = await chunkUpload(file, uploadConfig);
+                let response = await Axios.post(url,
+                    new FormData(), {
+                        params: { ...params, filename },
+                        proxy: config.proxy,
+                        headers: { 'Upload-Finish': true },
+                    });
 
                 return new Promise((resolve, reject) => {
-                    async function request() {
+                    async function checkStatus() {
                         try {
-                            const response = await Axios.post(`${backendAPI}/tasks/backup`, taskData, {
+                            taskData.set('rq_id', response.data.rq_id);
+                            response = await Axios.post(url, taskData, {
                                 proxy: config.proxy,
                                 params,
                             });
                             if (response.status === 202) {
-                                taskData = new FormData();
-                                taskData.append('rq_id', response.data.rq_id);
-                                setTimeout(request, 3000);
+                                setTimeout(checkStatus, 3000);
                             } else {
                                 // to be able to get the task after it was created, pass frozen params
                                 const importedTask = await getTasks({ id: response.data.id, ...params });
@@ -729,7 +754,7 @@
                         }
                     }
 
-                    setTimeout(request);
+                    setTimeout(checkStatus);
                 });
             }
 
@@ -766,19 +791,38 @@
                 // keep current default params to 'freeze" them during this request
                 const params = enableOrganization();
 
-                let data = new FormData();
-                data.append('project_file', file);
+                const projectData = new FormData();
+                const uploadConfig = {
+                    chunkSize: config.uploadChunkSize * 1024 * 1024,
+                    endpoint: `${origin}${backendAPI}/projects/backup/`,
+                    totalSentSize: 0,
+                    totalSize: file.size,
+                };
+
+                const url = `${backendAPI}/projects/backup`;
+                await Axios.post(url,
+                    new FormData(), {
+                        params,
+                        proxy: config.proxy,
+                        headers: { 'Upload-Start': true },
+                    });
+                const { filename } = await chunkUpload(file, uploadConfig);
+                let response = await Axios.post(url,
+                    new FormData(), {
+                        params: { ...params, filename },
+                        proxy: config.proxy,
+                        headers: { 'Upload-Finish': true },
+                    });
 
                 return new Promise((resolve, reject) => {
                     async function request() {
                         try {
-                            const response = await Axios.post(`${backendAPI}/projects/backup`, data, {
+                            projectData.set('rq_id', response.data.rq_id);
+                            response = await Axios.post(`${backendAPI}/projects/backup`, projectData, {
                                 proxy: config.proxy,
                                 params,
                             });
                             if (response.status === 202) {
-                                data = new FormData();
-                                data.append('rq_id', response.data.rq_id);
                                 setTimeout(request, 3000);
                             } else {
                                 // to be able to get the task after it was created, pass frozen params