+++ cvat/apps/engine/mixins.py
@@ -45,7 +45,12 @@ class TusFile:
         file_path = os.path.join(self.upload_dir, self.filename)
         file_exists = os.path.lexists(os.path.join(self.upload_dir, self.filename))
         if file_exists:
-            raise FileExistsError("File {} is already uploaded".format(self.filename))
+            original_file_name, extension = os.path.splitext(self.filename)
+            file_amount = 1
+            while os.path.lexists(os.path.join(self.upload_dir, self.filename)):
+                self.filename = "{}_{}{}".format(original_file_name, file_amount, extension)
+                file_path = os.path.join(self.upload_dir, self.filename)
+                file_amount += 1
         os.rename(file_id_path, file_path)
 
     def clean(self):
@@ -66,7 +71,8 @@ class TusFile:
     @staticmethod
     def create_file(metadata, file_size, upload_dir):
         file_id = str(uuid.uuid4())
-        cache.add("tus-uploads/{}/filename".format(file_id), "{}".format(metadata.get("filename")), TusFile._tus_cache_timeout)
+        filename = metadata.get("filename")
+        cache.add("tus-uploads/{}/filename".format(file_id), "{}".format(filename), TusFile._tus_cache_timeout)
         cache.add("tus-uploads/{}/file_size".format(file_id), file_size, TusFile._tus_cache_timeout)
         cache.add("tus-uploads/{}/offset".format(file_id), 0, TusFile._tus_cache_timeout)
         cache.add("tus-uploads/{}/metadata".format(file_id), metadata, TusFile._tus_cache_timeout)
@@ -175,7 +181,8 @@ class UploadMixin(object):
                 location = request.META.get('HTTP_ORIGIN') + request.META.get('PATH_INFO')
             return self._tus_response(
                 status=status.HTTP_201_CREATED,
-                extra_headers={'Location': '{}{}'.format(location, tus_file.file_id)})
+                extra_headers={'Location': '{}{}'.format(location, tus_file.file_id),
+                               'Upload-Filename': tus_file.filename})
 
     def append_tus_chunk(self, request, file_id):
         if request.method == 'HEAD':
@@ -202,7 +209,8 @@ class UploadMixin(object):
                 tus_file.clean()
 
             return self._tus_response(status=status.HTTP_204_NO_CONTENT,
-                                    extra_headers={'Upload-Offset': tus_file.offset})
+                                      extra_headers={'Upload-Offset': tus_file.offset,
+                                                     'Upload-Filename': tus_file.filename})
 
     def validate_filename(self, filename):
         upload_dir = self.get_upload_dir()